pyproject.toml
[tool.poetry]
name = "chat-interface"
version = "0.1.0"
description = "A ChatGPT-like interface for AI models"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.0.1"
python-dotenv = "^0.19.0"
requests = "^2.26.0"
litellm = "^1.30.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 